Add files via upload
Alright, so I added the .form and the .java, pretty sure the .java is more than enough, but I figured it's good to be safe. Everything should work, and my "change password" branch does check that the password meets the standards. To test with it just create a new EditAccountFrame and make it visible.